move packetout_test.go code, add the missing step related to entry programming and add Priority usage for ternary match
